Replace Finch by Mint (#2)
This change is needed to in order to avoid spawn a new process inside our processor,
and to remove the internal queue. Instead we are relying on Mint directly and using a
list only to keep the tweets from the last HTTP2 message.
After receiving the tweets we dispatch them right away.